Add files via upload
Completed assignment 4 for Harvard's CS50 Intro to Game Development Class. This assignment included creating a matching key and lock-block in every level. Once the player obtains the key and hits the lock, then a flag pole spawns at the end of the level. Once the player hits the pole then a new level begins which is 100 more tiles in width.